remove generic reader env, add specific Int# env to stateful
The issue is that if the generic env is not forced in the library, it's sometimes not unboxed in
user code, but if it's forced, then sometimes it's not removed from user code when it should be.

A potential solution would be to try using levity polymorphism tricks to have environments with
guaranteed unboxing and erasure.